Added backend authentication
