+++ .env
@@ -1,0 +1,1 @@
+MONGO_URI=mongodb+srv://restoadmin:queuingdb123@restaurantqueuing.7znjw.mongodb.net/queuing?retryWrites=true&w=majority
