+++ .gitignore
@@ -1,8 +1,1 @@
-# dependencies
-/node_modules
-
-.env
-
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
+node_modules
