--- README.md
@@ -1,7 +1,0 @@
-[![Work in Repl.it](https://classroom.github.com/assets/work-in-replit-14baed9a392b3a25080506f3b7b6d57f295ec2978f6f33ec97e36a161684cbe9.svg)](https://classroom.github.com/online_ide?assignment_repo_id=294623&assignment_repo_type=GroupAssignmentRepo)
-# sd-lab-project-team-v-cube
-sd-lab-project-team-v-cube created by GitHub Classroom
-
------ 
-#### Topic for project
-### Restaurant queuing application
