+++ client/.gitignore
@@ -1,0 +1,8 @@
+# dependencies
+/node_modules
+
+.env
+
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
