+++ client/src/components/Login/Login.css
@@ -1,0 +1,76 @@
+#login-main {
+  width: 100vw;
+  height: 100vh;
+  background-image: url('../../images/loginBackground.jpg');
+}
+
+#login-page {
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  margin-top: 350px;
+  margin-left: auto;
+  margin-right: auto;
+  width: 270px;
+  height: 330px;
+  color: black;
+  padding: 30px;
+  border-radius: 10px;
+  border: 1px solid orange;
+  border-radius: 5px;
+  background: white;
+}
+
+#login-img {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  margin-top: 100px;
+  margin-left: auto;
+  margin-right: auto;
+  height: 600px;
+  width: 320px;
+  box-shadow: 0px 0px 4px 4px #555;
+  border-radius: 5px;
+}
+
+@media (max-width:350px) {
+  #login-page {
+    width: 250px;
+    height: 280px;
+  }
+
+  #login-img {
+    height: 0;
+    width: 0;
+  }
+}
+
+.login-btn:hover {
+  text-decoration: none;
+  background: white;
+}
+
+#password {
+  opacity: 1;
+}
+
+#login-button {
+  display: block;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+#login-button:hover {
+  background: none;
+  border: 1px solid orange;
+  color: black;
+}
+
+#login-text {
+  text-align: center;
+}
