+++ client/src/images/logo.png
