+++ models/Booking.js
@@ -1,0 +1,11 @@
+const mongoose = require('mongoose');
+
+const bookingSchema = new mongoose.Schema({
+    _id: Number,
+    size: Number,
+    time: String,
+    mobile: String,
+    email: String
+})
+
+module.exports = mongoose.model('Booking', bookingSchema);
