+++ models/User.js
@@ -1,0 +1,12 @@
+const mongoose = require('mongoose');
+const passportLocalMongoose = require('passport-local-mongoose');
+
+const userSchema = new mongoose.Schema({
+    username: String,
+    password: String,
+    flag: String,
+    resto: String
+})
+
+userSchema.plugin(passportLocalMongoose);
+module.exports = mongoose.model('User', userSchema);
