+++ package.json
@@ -1,49 +1,37 @@
 {
-  "name": "restaurant-queuing",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "bcryptjs": "^2.4.3",
-    "body-parser": "^1.19.0",
-    "bootstrap": "^4.3.1",
-    "concurrently": "^5.0.2",
-    "express": "^4.17.1",
-    "express-session": "^1.17.0",
-    "is-empty": "^1.2.0",
-    "jsonwebtoken": "^8.5.1",
-    "mongodb": "^3.3.2",
-    "mongoose": "^5.8.9",
-    "passport": "^0.4.1",
-    "passport-jwt": "^4.0.0",
-    "passport-local": "^1.0.0",
-    "passport-local-mongoose": "^6.0.1",
-    "re-base": "^4.0.0",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
-    "react-router-dom": "^5.0.1",
-    "react-scripts": "3.1.1",
-    "reactstrap": "^8.0.1",
-    "validator": "^11.1.0"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  }
+	"name": "hackathon",
+	"version": "1.0.0",
+	"description": "",
+	"main": "server.js",
+	"scripts": {
+		"test": "echo \"Error: no test specified\" && exit 1",
+		"server": "nodemon server.js",
+		"start": "node server.js",
+		"client": "npm start --prefix client",
+		"dev": "concurrently \"npm run server\" \"npm run client\"",
+		"heroku-postbuild": "cd client && npm install --only=dev && npm install && npm run build"
+	},
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
+	"proxy": "http://localhost:5000/",
+	"dependencies": {
+		"body-parser": "^1.19.0",
+		"connect-mongo": "^3.2.0",
+		"cookie-parser": "^1.4.5",
+		"cors": "^2.8.5",
+		"csv-parser": "*",
+		"dotenv": "^8.2.0",
+		"express": "^4.17.1",
+		"express-session": "^1.17.1",
+		"mongoose": "^5.7.9",
+		"nodemailer": "^6.4.6",
+		"passport": "^0.4.1",
+		"passport-local-mongoose": "^6.0.1",
+		"sentiment": "^5.0.2"
+	},
+	"devDependencies": {
+		"concurrently": "^5.3.0",
+		"nodemon": "^1.19.4"
+	}
 }
